Serve service worker JS from / so it can control the root scope
